Moved Typing to Message component
