+++ gtkcord/components/message/typing/typing.go
@@ -84,8 +84,6 @@ type State struct {
 	*gtk.Box
 	Label *gtk.Label
 
-	ChannelID discord.Snowflake
-
 	mu sync.Mutex
 
 	Users []typingUser
